// src/types/oldDb/vehicle.ts
export interface IOldVehicle {
  id: number;
  main_photo_id: number;
  user_id: number;
  date_added: string;
  kind_pic: number;
  name: string;
  photos: number;
  kind: number;
  kind_id: number;
  youtube: string;
  txt: string;
  brand_id: number;
  model_id: number;
  year: number;
  amortization: number;
  wheels: number;
  brake_id: number;
  weight: number;
  speed: number;
  driver_id: number;
  voltage: number;
  watage: number;
  engine_id: number;
  engine_brand_id: number;
  engine_model_id: number;
  engine_winding: "";
  battery_install_id: number;
  accumulator_id: number;
  serial_connection: number;
  parallel_connection: number;
  battery_voltage: number;
  wph: number;
  aph: number;
  cell_brand_id: number;
  cell_model_id: number;
  likes: number;
  views: number;
  power: number;
  bike_range: number;
  city: string;
  del_val: number;
}